created debug.py
- file contains all debug functions
- added now function for adjusting the threshold live
